requirements.txt
fastapi==0.115.14
faster-whisper== 1.1.1
uvicorn==0.35.0
python-dotenv==1.1.1
langchain-groq==0.3.5
torch==2.7.1
torchaudio==2.7.1
pyannote.audio==3.3.2
reportlab==4.4.2